Fix some FK datatypes (unsignedInt to unsignedBigInt) and string for formateur matricule FK, migrations order (formateurs above module) and add formateur_id to module seeder

+++ backend/database/migrations/2024_03_10_114855_create_groupes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string("name");
             $table->string("code");
             $table->enum("annee",array(["1A","2A","3A"]));
-            $table->unsignedInteger("filiere_id");
+            $table->unsignedBigInteger("filiere_id");
             $table->timestamps();
             
             $table->foreign('filiere_id')->references('id')->on('filieres')->onDelete('cascade');
